polymarket-mm/requirements.txt
pymongo==4.6.1
websocket-client==1.7.0
websockets==12.0
python-dotenv==1.0.0
requests==2.31.0
py-clob-client==0.26.0
asyncio-mqtt==0.11.0
pydantic==2.5.3